First Function
It seems apt that the first custom function published to this repository
should also be our example function that ships with the product.
Should now be able to pull, compile and go (though may need to edit the
plugin.xml first as that is set up in anticipation of my cleaning up all
the other functions).