Updated to be based on a Maven build approach; this should be development environment agnostic.
This means the dependencies are now managed by Maven as is the build and packaging into a Jar file.
I have removed the TelnetConnection example as the library I was using for this is not available in Maven and I don't have time to re-write it at the moment (excersize for the user to build a TelnetConnection example using Apache Commons Net).

Please see the readme for build notes as this update does require some additional configuration due to being unable to ship proprietary libraries with the functions.

Signed-off-by: chaggan <chaggan@gmail.com>